requirements.txt
django

psycopg
psycopg[binary,pool]

python-dotenv

djangorestframework
django-cors-headers